parcial_2/practica_06a_datos_normales/codigo/python/requirements.txt
paho-mqtt
pandas
scikit-learn
matplotlib
joblib
